move set statement to dcl
